requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
click==7.1.2
django-debug-toolbar==3.2.1
django-ninja==0.13.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8.2
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0